이미지 업로드 기능 구현 및 UI, 테스트 완료

+++ admin/src/app/actions/utApi.ts
@@ -1,10 +1,18 @@
 "use server";
 
+import { NextResponse } from "next/server";
+import { toast } from "sonner";
 import { UTApi } from "uploadthing/server";
 
 export const utApi = new UTApi();
 
 export async function deleteImages(urls: string[]) {
+  if (Array.isArray(urls) && urls.length) {
+    toast.warning("적합한 URL 타입이 아닙니다.");
+    
+    return NextResponse.json({ error: "Invalid urls type." }, { status: 400 });
+  }
+
   try {
     await utApi.deleteFiles(urls);
   } catch (err) {
